integration: add corruption test
